+++ src/main/java/cloud/glitchdev/rfu/access/PlhlegblastStateAccess.kt
@@ -1,0 +1,6 @@
+package cloud.glitchdev.rfu.access
+
+interface PlhlegblastStateAccess {
+    fun `rfu$isPlhlegblast`(): Boolean
+    fun `rfu$setPlhlegblast`(isPlhlegblast: Boolean)
+}